Fix comparison type mismatch in getdents64 test helper
The `count_using_getdents64_fast` function within the `file_count_tests` module failed to compile on certain platforms or compiler versions due to a type mismatch error (E0308/E0277).

The error occurred when comparing dereferenced byte values from the `d_name` pointer (which the compiler treated as `u8`) against byte literals (`b'.'`) that were explicitly cast to `i8`. This resulted in an invalid attempt to compare `u8` with `i8`.

This commit removes the unnecessary `as i8` casts from the comparisons involving `b'.'`. This means the comparison is directly between `u8` values (the dereferenced byte and the byte literal), resolving the compilation error and making the code more robust across different environments.

+++ src/main.rs
@@ -2683,7 +2683,7 @@ mod file_count_tests {
                     let reclen = (*d).d_reclen as usize;
                     let name_ptr = (*d).d_name.as_ptr();
                     // Fast manual check for "." and ".." without creating a CStr.
-                    if *name_ptr != b'.' as i8 {
+                    if *name_ptr != b'.' {
                         if (*d).d_type == libc::DT_REG {
                             count += 1;
                         }
@@ -2691,7 +2691,7 @@ mod file_count_tests {
                         let second = *name_ptr.add(1);
                         if second == 0 {
                             // It's ".", skip.
-                        } else if second == b'.' as i8 && *name_ptr.add(2) == 0 {
+                        } else if second == b'.' && *name_ptr.add(2) == 0 {
                             // It's "..", skip.
                         } else if (*d).d_type == libc::DT_REG {
                             count += 1;
